Ajuste na validação de dados e regra de negócio da exclusão de empresa, deleteCompany

+++ src/modules/companies/useCases/deleteCompany/deleteCompanyService.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from "tsyringe";
+import { AppError } from "../../../../errors/appError";
 import { ICompanyRepository } from "../../repositories/iCompanyRepository";
 
 @injectable()
@@ -6,6 +7,12 @@ export class DeleteCompanyService {
   constructor(@inject("CompanyRepository") private companyRepository: ICompanyRepository) { }
 
   async execute(id: string): Promise<void> {
+    const idExists = await this.companyRepository.findIdCompany(id);
+
+    if (!idExists) {
+      throw new AppError("Não existe uma empresa com este ID cadastrado");
+    }
+
     await this.companyRepository.deleteCompany(id);
   }
 }
